Add ability to log in/out
This will change the login page to actually allow log in and remember
users. The navbar will also change site-wide for logged in users,
showing their account and a logout button in the top left.

This also updates login to use our Account model class, and updates
the Account class to allow creation from an Account database object.