Add some game finished logic for lobby games
This implements most of the game finished logic for lobby games (and
implementing it for hotseat/ai will be trivial now). All that still
needs to be done is set up the lobby controller to handle it (i.e. block
the losers, wait for all the users before continuing, etc.) which
shouldn't be too difficult.