Finish game pages
This finishes up all the lobby and game controllers, and adds the AI
quick game page and improves the hotseat view.

Note that because the Game class isn't calculating territories properly
right now, the AI will stop working once the board reaches an invalid
state.

Once the Game class is set up, the app will be pretty much good enough
to present. There is still some kinda broken stuff, but overall if we
run out of time it should be good enough.